isp-crm-system/isp-crm-system/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
httpx==0.27.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery[redis]==5.4.0
redis==5.0.8
python-dotenv==1.0.1
